requirements.txt
appdirs==1.4.4
base58==2.1.0
black==20.8b1
click==7.1.2
fastapi==0.63.0
h11==0.12.0
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8
pymultihash==0.8.2
python-multipart==0.0.5
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
